moved item stats view to the stats_views file
